Clean and move decorators
Clean unused decorators
Added CORS
Move measure to a separate file
Fix tests
