+++ kilink/kilink.py
@@ -4,9 +4,6 @@
 """The server for kilink."""
 
 import logging
-import time
-
-from functools import update_wrapper
 
 from flask import (
     Flask,
@@ -16,8 +13,7 @@ from flask import (
     make_response
 )
 
-# from flask.ext.assets import Environment
-# from flask_assets import Environment
+from flask_cors import CORS
 from flask_babel import Babel
 from sqlalchemy import create_engine
 
@@ -25,8 +21,7 @@ import backend
 import loghelper
 
 from config import config, LANGUAGES
-from metrics import StatsdClient
-from decorators import crossdomain
+from decorators import measure
 
 # set up flask
 app = Flask(__name__)
@@ -36,55 +31,10 @@ app.config["STATIC_ROOT"] = 'static'
 app.config["PROPAGATE_EXCEPTIONS"] = False
 
 babel = Babel(app)
-
-# flask-assets
-# assets = Environment(app)
-# assets.cache = "/tmp/"
-# assets.init_app(app)
+cors = CORS(app)
 
 # logger
 logger = logging.getLogger('kilink.kilink')
-
-# metrics
-metrics = StatsdClient("linkode")
-
-
-def nocache(f):
-    """Decorator to make a page un-cacheable."""
-    def new_func(*args, **kwargs):
-        """The new function."""
-        resp = make_response(f(*args, **kwargs))
-        resp.headers['Cache-Control'] = 'public, max-age=0'
-        return resp
-    return update_wrapper(new_func, f)
-
-
-def measure(metric_name):
-    """Decorator generator to send metrics counting and with timing."""
-
-    def _decorator(oldf):
-        """The decorator itself."""
-
-        def newf(*args, **kwargs):
-            """The function to replace."""
-            tini = time.time()
-            try:
-                result = oldf(*args, **kwargs)
-            except Exception as exc:
-                name = "%s.error.%s" % (metric_name, exc.__class__.__name__)
-                metrics.count(name, 1)
-                raise
-            else:
-                tdelta = time.time() - tini
-
-                metrics.count(metric_name + '.ok', 1)
-                metrics.timing(metric_name, tdelta)
-                return result
-
-        # need to fix the name because it's used by flask
-        newf.func_name = oldf.func_name
-        return newf
-    return _decorator
 
 
 @app.errorhandler(backend.KilinkNotFoundError)
@@ -107,7 +57,7 @@ def get_locale():
     return request.accept_languages.best_match(LANGUAGES.keys())
 
 
-# accesory pages
+# accessory pages
 @app.route('/about')
 @measure("about")
 def about():
@@ -160,7 +110,6 @@ def index(linkode_id=None, revno=None):
 
 # API
 @app.route('/api/1/linkodes/', methods=['POST'])
-@crossdomain(origin='*')
 @measure("api.create")
 def api_create():
     """Create a kilink."""
@@ -181,7 +130,6 @@ def api_create():
 
 
 @app.route('/api/1/linkodes/<linkode_id>', methods=['POST'])
-@crossdomain(origin='*')
 @measure("api.update")
 def api_update(linkode_id):
     """Update a kilink."""
@@ -210,7 +158,6 @@ def api_update(linkode_id):
 
 @app.route('/api/1/linkodes/<linkode_id>/<revno>', methods=['GET'])
 @app.route('/api/1/linkodes/<linkode_id>', methods=['GET'])
-@crossdomain(origin='*')
 @measure("api.get")
 def api_get(linkode_id, revno=None):
     """Get the kilink and revno content"""
